improve neon code.
